Added some funtionallity and improved the code
-Added a JList to display the online users. Using this we can let the
user click on the user he want to have a prive conversation instead of
having to type the whole name of the user. The list is synced with
usersList on the clientside - the online users is displayed.
BUT: the JList seems to not be fully syned with the ArrayList
"usersList". I think the error is caused by multithreading.
-Improved the overall code.